Update clean.ps1
Removing everything with 'bin', 'obj' or 'packages' through the whole directory structure makes a few assumptions that those folders aren't present anywhere else... For example I ran this on the actual NuGetGallery solution, which has ASP.NET MVC Views/Packages area, so in that case did not want to delete Packages. Ideally, just want to delete bin / obj / packages folders on the same level as a .sln or a .csproj or .*proj file types.